Seperating the watch managment and watch face, so we can have multiple watch faces
