+++ src/WatchFace.h
@@ -1,0 +1,6 @@
+#pragma once
+
+class CWatchFace
+{
+
+};
