+++ src/WatchyExpanded.h
@@ -1,90 +1,27 @@
 #pragma once
 
-#include <Arduino.h>
-#include <WiFiManager.h>
-#include <HTTPClient.h>
-#include <NTPClient.h>
-#include <WiFiUdp.h>
-#include <Arduino_JSON.h>
-#include <GxEPD2_BW.h>
-#include <Wire.h>
-#include <Fonts/FreeMonoBold9pt7b.h>
-#include "DSEG7_Classic_Bold_53.h"
-#include "WatchyRTC.h"
-#include "BLE.h"
-#include "bma.h"
-#include "config.h"
+// STL
+#include <vector.h>
 
-typedef struct weatherData{
-	int8_t temperature;
-	int16_t weatherConditionCode;
-	bool isMetric;
-	String weatherDescription;
-}weatherData;
+// Defs
+class CWatchFace;
 
-typedef struct watchySettings{
-	//Weather Settings
-	String cityID;
-	String weatherAPIKey;
-	String weatherURL;
-	String weatherUnit;
-	String weatherLang;
-	int8_t weatherUpdateInterval;
-	//NTP Settings
-	String ntpServer;
-	int gmtOffset;
-	int dstOffset;
-} watchySettings;
-
-class WatchyExpanded
+class CWatchyExpanded
 {
 	public:
-		static WatchyRTC RTC;
-		static GxEPD2_BW<GxEPD2_154_D67, GxEPD2_154_D67::HEIGHT> display;
-		tmElements_t currentTime;
-		watchySettings settings;
-	public:
-		explicit WatchyExpanded(const watchySettings& s) : settings(s){} //constructor
-		void init(String datetime = "");
-		void deepSleep();
-		static void displayBusyCallback(const void*);
-		float getBatteryVoltage();
-		void vibMotor(uint8_t intervalMs = 100, uint8_t length = 20);
+		CWatchyExpanded();
 
-		void handleButtonPress();
-		void showMenu(byte menuIndex, bool partialRefresh);
-		void showFastMenu(byte menuIndex);
-		void showAbout();
-		void showBuzz();
-		void showAccelerometer();
-		void showUpdateFW();
-		void showSyncNTP();
-		bool syncNTP();
-		bool syncNTP(long gmt, int dst, String ntpServer);
-		void setTime();
-		void setupWifi();
-		bool connectWiFi();
-		weatherData getWeatherData();
-		weatherData getWeatherData(String cityID, String units, String lang, String url, String apiKey, uint8_t updateInterval);
-		void updateFWBegin();
+		void AddWatchFace(CWatchFace* pFace);
 
-		void showWatchFace(bool partialRefresh);
-		virtual void drawWatchFace(); //override this method for different watch faces
-
-		// Expanded
-		JSONVar& getWeatherJSON();
-		JSONVar& getWeatherJSON(String cityID, String units, String lang, String url, String apiKey,
-		                        uint8_t updateInterval);
+		void Init();
 
 	private:
-		void _bmaConfig();
-		static void _configModeCallback(WiFiManager *myWiFiManager);
-		static uint16_t _readRegister(uint8_t address, uint8_t reg, uint8_t *data, uint16_t len);
-		static uint16_t _writeRegister(uint8_t address, uint8_t reg, uint8_t *data, uint16_t len);
-};
+		void displayBusyCallback(const void*);
+		void UpdateScreen();
 
-extern RTC_DATA_ATTR int guiState;
-extern RTC_DATA_ATTR int menuIndex;
-extern RTC_DATA_ATTR BMA423 sensor;
-extern RTC_DATA_ATTR bool WIFI_CONFIGURED;
-extern RTC_DATA_ATTR bool BLE_CONFIGURED;
+		std::vector<CWatchFace*> m_faces;
+		std::int8_t m_face = 0;
+
+		GxEPD2_BW<GxEPD2_154_D67, GxEPD2_154_D67::HEIGHT> m_display;
+		tmElements_t m_currentTime;
+};
